feat: add landing and template pages
